lzjb_compress done.
the key of lzjb compress is finding the longest match in previous scaned data:
it has some restricts on offset and match length because of only 10 and 6 bits
for them each.

efficiency can be improved by optimizing match searching. hash the next 3 chars
to address pointer if they occurred in scaned data, caching the preivous matching
results, will help a lot. But we use a exhault method here.